Assorted fixes.
