+++ client/src/cards/data.ts
@@ -4,6 +4,7 @@ import {thoriumPath} from "server/src/utils/appPaths";
 import type {FlightDataModel} from "server/src/classes/FlightDataModel";
 import {Entity} from "server/src/utils/ecs";
 import {parse} from "yaml";
+import {getFlights} from "server/src/utils/getFlights";
 // This file is used for any subscriptions which all clients
 // make, regardless of what cards they have.
 export const subscriptions = {
@@ -13,7 +14,8 @@ export const subscriptions = {
     // TODO Aug 31, 2021 - Merge this data with the flight client data
     // or create a new subscription just for the flight client.
     const {id, name, connected} = context.server.clients[context.clientId];
-    return {id, name, connected};
+
+    return {id, name, connected, ...context.flightClient};
   },
   flight(context: DataContext) {
     const flight = context.flight;
@@ -22,25 +24,7 @@ export const subscriptions = {
     return {date, name, paused};
   },
   flights: async (context: DataContext) => {
-    let files: string[];
-    try {
-      files = await fs.readdir(`${thoriumPath}/flights/`);
-    } catch {
-      await fs.mkdir(`${thoriumPath}/flights/`);
-      files = [];
-    }
-    const flightFiles = files.filter(f => f.includes(".flight"));
-    const flightData = await Promise.all(
-      flightFiles.map(async flightName => {
-        const raw = await fs.readFile(
-          `${thoriumPath}/flights/${flightName}`,
-          "utf-8"
-        );
-        const data = parse(raw);
-        return {...data, date: new Date(data.date)} as FlightDataModel;
-      })
-    );
-    return flightData;
+    return getFlights();
   },
   dots(context: DataContext) {
     return (
